requirements.txt
APScheduler==3.10.4
faiss_cpu==1.10.0
fastapi==0.115.12
numpy==2.2.4
pydantic==2.11.1
Requests==2.32.3
sentence_transformers==3.4.1
uvicorn==0.34.0
watchdog==4.0.2
